applied suggestions - refactored validate_iso_week method to common util
