+++ smart_tests/commands/view/flaky_tests.py
@@ -1,5 +1,4 @@
 import json
-import re
 import sys
 from http import HTTPStatus
 from typing import Annotated
@@ -8,30 +7,11 @@ import click
 
 import smart_tests.args4p.typer as typer
 from smart_tests.args4p.converters import intType
-from smart_tests.args4p.exceptions import BadCmdLineException
 
 from ... import args4p
 from ...app import Application
 from ...utils.smart_tests_client import SmartTestsClient
-from ...utils.typer_types import DateTimeWithTimezone, parse_datetime_with_timezone
-
-
-def validate_iso_week(value: str) -> str:
-    """Validate ISO week format YYYY-Www and week number range (01-53)"""
-    pattern = r'^\d{4}-W\d{2}$'
-    if not re.match(pattern, value):
-        raise BadCmdLineException(
-            f"Invalid year-week format: '{value}'. Expected format: YYYY-Www (e.g., 2026-W15)"
-        )
-
-    # Extract and validate week number range
-    week_num = int(value.split('-W')[1])
-    if week_num < 1 or week_num > 53:
-        raise BadCmdLineException(
-            f"Invalid week number: '{value}'. Week must be between 01 and 53"
-        )
-
-    return value
+from ...utils.typer_types import DateTimeWithTimezone, parse_datetime_with_timezone, validate_iso_week
 
 
 @args4p.command(help="View flaky test data with weekly scores")
